Update deleteListing.php
So this is the only thing i have changed tonight to see if it works but it showed that their is either an issue in my DB function $stament executable or that im just testing it wrong and everything may be good. Can u just look over the DB functions i had from a few days ago in make sure they are good. I tried it with the db function deleteListingByListingIDAndUserID, and i tried it with deleteListingByListingID and i had the same issue come up